Move ProfileV3APIList to v3 directory
