+++ app/src/main/java/nl/eduvpn/app/entity/v3/Info.kt
@@ -2,7 +2,6 @@ package nl.eduvpn.app.entity.v3
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import nl.eduvpn.app.entity.ProfileV3APIList
 
 @Serializable
 data class Info(
